requirements.txt
azure-storage-blob==1.1.0
